# atividade-13/jwt-auth-service/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          }
        },
        "summary": "Registra um novo usuário",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "summary": "Realiza login com email e senha",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/perfil": {
      "get": {
        "operationId": "AuthController_getPerfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil retornado com sucesso"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retorna perfil do usuário autenticado",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin": {
      "get": {
        "operationId": "AdminController_getAdminData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Acesso autorizado para ADMIN"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Endpoint de administração (apenas ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/user/perfil": {
      "get": {
        "operationId": "UserController_getUserPerfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil retornado com sucesso"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Perfil do usuário (somente USER)",
        "tags": [
          "user"
        ]
      }
    }
  },
  "info": {
    "title": "API com Swagger",
    "description": "Documentação automática da API com Swagger",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "senhaSegura123",
            "minLength": 6
          },
          "name": {
            "type": "string",
            "example": "João Silva"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "senhaSegura123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}